# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: JianArtifact API
  version: 0.4.0
  description: 自托管多格式制品仓库的管理 API。协议端点（Raw/Maven/npm 等原生客户端交互）以协议兼容测试保障，不在此契约内描述。
  license:
    name: MIT
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/v1/observability/downloads/by-client:
    get:
      operationId: getDownloadByClient
      summary: 制品下载来源聚合（仅管理员）
      description: |
        按来源 IP（Top 10，明文仅管理员）与客户端类型（UA 归类，不暴露原始 UA 串）聚合。
        去重口径固定为「同 IP + 同制品 1 小时窗口只计一次贡献」（独立来源视角）；
        原始累计口径见仪表盘 downloadTrend。
      parameters:
        - $ref: "#/components/parameters/ObservabilityFromParam"
        - $ref: "#/components/parameters/ObservabilityToParam"
      responses:
        "200":
          description: 下载来源聚合
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadClientRanking"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/observability/downloads/trend:
    get:
      operationId: getDownloadTrendGrouped
      summary: 分组下载趋势与分组总计（仅管理员）
      description: |
        按桶输出分组下载时序 `points`（稀疏「桶 × 组」，缺失桶即 0，前端对齐桶轴补零）
        与全窗口分组总计 `totals`（按计数降序，饼图数据源）。
        groupBy=family 按 UA 归类分组（默认，不暴露原始 UA 串）；groupBy=ip 返回来源 IP 明文，
        与 by-client 同一隐私边界（IP 明文仅管理员可见，端点整体仅管理员）。
        采集口径与仪表盘 downloadTrend 同源（原始累计）；from/to 成对提供或全省略（缺省最近 24 小时），跨度最长 30 天。
      parameters:
        - $ref: "#/components/parameters/ObservabilityFromParam"
        - $ref: "#/components/parameters/ObservabilityToParam"
        - $ref: "#/components/parameters/DownloadGroupByParam"
        - $ref: "#/components/parameters/DownloadRepoParam"
      responses:
        "200":
          description: 分组下载趋势（桶轴参数与稀疏点同响应返回）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadGroupedTrendResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/observability/dashboard:
    get:
      operationId: getOperationsDashboard
      summary: 当前节点业务仪表盘（仅管理员）
      description: 只返回当前实例的真实协议聚合和逻辑制品容量，不含主机或其他节点数据。
      parameters:
        - $ref: "#/components/parameters/ObservabilityFromParam"
        - $ref: "#/components/parameters/ObservabilityToParam"
      responses:
        "200":
          description: 当前节点业务概览
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationsDashboard"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/observability/host:
    get:
      operationId: getHostMonitoring
      summary: 当前主机监控（仅管理员）
      description: 只返回当前实例所在主机的持久化分钟样本，不读取或汇总其他节点。
      parameters:
        - $ref: "#/components/parameters/ObservabilityFromParam"
        - $ref: "#/components/parameters/ObservabilityToParam"
      responses:
        "200":
          description: 当前主机概览和趋势
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostMonitoring"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ObservabilityFromParam:
      name: from
      in: query
      required: false
      description: UTC 时间范围下界（含）。与 to 同时提供或同时省略；省略时为最近 24 小时。
      schema:
        type: string
        format: date-time
    ObservabilityToParam:
      name: to
      in: query
      required: false
      description: UTC 时间范围上界（不含），最长 30 天。
      schema:
        type: string
        format: date-time
    DownloadGroupByParam:
      name: groupBy
      in: query
      required: false
      description: 分组维度：family（缺省，UA 归类）或 ip（来源 IP 明文，仅管理员）；非法值返回 400。
      schema:
        type: string
        enum: [family, ip]
        default: family
    DownloadRepoParam:
      name: repo
      in: query
      required: false
      description: 按仓库名精确过滤；省略为全部仓库。
      schema:
        type: string
        minLength: 1
        maxLength: 255
...
  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 越权
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Conflict:
      description: 冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    ObservabilityBucket:
      type: string
      enum: [minute, hour, day]
    ProtocolMetricPoint:
      type: object
      required: [from, to, requestCount, downloadCount, failureCount, cacheHitCount, cacheMissCount]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        requestCount:
          type: integer
          format: int64
          minimum: 0
        downloadCount:
          type: integer
          format: int64
          minimum: 0
        failureCount:
          type: integer
          format: int64
          minimum: 0
        cacheHitCount:
          type: integer
          format: int64
          minimum: 0
        cacheMissCount:
          type: integer
          format: int64
          minimum: 0
    CapacityPoint:
      type: object
      required: [from, to, repositoryCount, assetCount, logicalBytes]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        repositoryCount:
          type: integer
          format: int64
          minimum: 0
        assetCount:
          type: integer
          format: int64
          minimum: 0
        logicalBytes:
          type: integer
          format: int64
          minimum: 0
    OperationsAlert:
      type: object
      required: [code, severity, source, observedAt]
      properties:
        code:
          type: string
        severity:
          type: string
          enum: [warning, critical]
        source:
          type: string
        observedAt:
          type: string
          format: date-time
        # v0.8.0：告警按 (code, source) 去重持久化后，首次观察到的时间。
        # 未持久化（或首次即最新）时缺省；前端用于展示“持续中”的起始点。
        firstObservedAt:
          type: string
          format: date-time
        # v0.8.0：upstream_auto_blocked 的自动阻止窗口截止时间；其他告警缺省。
        blockedUntil:
          type: string
          format: date-time
    DownloadTrendPoint:
      type: object
      required: [from, to, downloadCount]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        downloadCount:
          type: integer
          format: int64
    DownloadIpCount:
      type: object
      required: [ip, count]
      properties:
        ip:
          type: string
        count:
          type: integer
          format: int64
    DownloadFamilyCount:
      type: object
      required: [family, count]
      properties:
        family:
          type: string
        count:
          type: integer
          format: int64
    DownloadClientRanking:
      type: object
      required: [topIps, families]
      properties:
        topIps:
          type: array
          items:
            $ref: "#/components/schemas/DownloadIpCount"
        families:
          type: array
          items:
            $ref: "#/components/schemas/DownloadFamilyCount"
    DownloadGroupedTrendPoint:
      type: object
      description: |
        分组下载时序的一点：稀疏「桶 × 组」记录，缺失桶即 0（服务端不按组做笛卡尔补零，
        前端对齐桶轴后把缺失桶记 0）。桶区间 [from, to) 左闭右开，from 即绘图 label。
      required: [from, to, group, count]
      properties:
        from:
          type: string
          format: date-time
          description: 桶起点（左闭），前端绘图 label 取此时间戳。
        to:
          type: string
          format: date-time
          description: 桶终点（右开）。
        group:
          type: string
          description: 分组取值：groupBy=ip 时为来源 IP 明文，=family 时为 UA 归类结果。
          minLength: 1
          maxLength: 255
        count:
          type: integer
          format: int64
          minimum: 0
    DownloadGroupTotal:
      type: object
      description: 单个分组在整段窗口内的累计下载（饼图数据源；按 count 降序返回）。
      required: [group, count]
      properties:
        group:
          type: string
          description: 分组取值：groupBy=ip 时为来源 IP 明文，=family 时为 UA 归类结果。
          minLength: 1
          maxLength: 255
        count:
          type: integer
          format: int64
          minimum: 0
    DownloadGroupedTrendResponse:
      type: object
      description: |
        分组下载趋势响应：points 为稀疏「桶 × 组」时序，totals 为全窗口按组降序总计；
        from/to/effectiveBucket 是桶轴参数，前端据此对齐缺失桶。
      required: [from, to, effectiveBucket, groupBy, points, totals]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        effectiveBucket:
          $ref: "#/components/schemas/ObservabilityBucket"
        groupBy:
          type: string
          description: 实际生效的分组维度（family 或 ip）。
          enum: [family, ip]
        points:
          type: array
          items:
            $ref: "#/components/schemas/DownloadGroupedTrendPoint"
        totals:
          type: array
          items:
            $ref: "#/components/schemas/DownloadGroupTotal"
...
    OperationsDashboard:
      type: object
      required:
        [
          from,
          to,
          effectiveBucket,
          current,
          kpi,
          requestTrend,
          downloadTrend,
          capacityTrend,
          alerts,
        ]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        effectiveBucket:
          $ref: "#/components/schemas/ObservabilityBucket"
        current:
          $ref: "#/components/schemas/CapacityPoint"
        kpi:
          $ref: "#/components/schemas/OperationsDashboardKpi"
        requestTrend:
          type: array
          items:
            $ref: "#/components/schemas/ProtocolMetricPoint"
        downloadTrend:
          type: array
          items:
            $ref: "#/components/schemas/DownloadTrendPoint"
        capacityTrend:
          type: array
          items:
            $ref: "#/components/schemas/CapacityPoint"
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/OperationsAlert"
    OperationsDashboardKpi:
      type: object
      description: 服务端按当前范围聚合的权威仪表盘 KPI；客户端不得自行从趋势重新计算。
      required:
        [
          repositoryCount,
          assetCount,
          logicalBytes,
          requestCount,
          downloadCount,
          failureCount,
          cacheHitRate,
        ]
      properties:
        repositoryCount:
          type: integer
          format: int64
          minimum: 0
        assetCount:
          type: integer
          format: int64
          minimum: 0
        logicalBytes:
          type: integer
          format: int64
          minimum: 0
        requestCount:
          type: integer
          format: int64
          minimum: 0
        downloadCount:
          type: integer
          format: int64
          minimum: 0
        failureCount:
          type: integer
          format: int64
          minimum: 0
        cacheHitRate:
          type: number
          format: double
          nullable: true
          minimum: 0
          maximum: 1
          description: 命中数除以命中与未命中样本数；无缓存样本时为 null。
    MetricGroupState:
      type: string
      enum: [ok, unavailable, unsupported, error]
    HostMetricGroup:
      type: object
      required: [state]
      properties:
        state:
          $ref: "#/components/schemas/MetricGroupState"
        errorCode:
          type: string
    HostMetricPoint:
      type: object
      required: [from, to, hostState, networkState, processState, readinessState]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        hostState:
          $ref: "#/components/schemas/HostMetricGroup"
        networkState:
          $ref: "#/components/schemas/HostMetricGroup"
        processState:
          $ref: "#/components/schemas/HostMetricGroup"
        readinessState:
          $ref: "#/components/schemas/HostMetricGroup"
        cpuPercent:
          type: number
          format: double
          nullable: true
        memoryTotalBytes:
          type: integer
          format: int64
          nullable: true
        memoryAvailableBytes:
          type: integer
          format: int64
          nullable: true
        # 同一采样点的 memoryTotalBytes − memoryAvailableBytes，固定口径，前端只展示不复算。
        memoryUsedBytes:
          type: integer
          format: int64
          nullable: true
          description: 内存已用（字节），= total − available，采样点固定口径。
        # 数据目录所在卷，与 diskAvailableBytes 同一口径（同一文件系统统计）。
        diskTotalBytes:
          type: integer
          format: int64
          nullable: true
          description: 磁盘总量（字节），数据目录所在卷，与 diskAvailableBytes 同口径。
        diskAvailableBytes:
          type: integer
          format: int64
          nullable: true
        # 同一采样点的 diskTotalBytes − diskAvailableBytes，与内存已用算法对称。
        diskUsedBytes:
          type: integer
          format: int64
          nullable: true
          description: 磁盘已用（字节），= total − available。
        networkReceiveBytesPerSecond:
          type: number
          format: double
          nullable: true
        networkTransmitBytesPerSecond:
          type: number
          format: double
          nullable: true
        processRssBytes:
          type: integer
          format: int64
          nullable: true
        processCpuPercent:
          type: number
          format: double
          nullable: true
        # 挂钟口径：仅统计本进程自身，不做系统进程枚举（与 processRssBytes 同一主体）。
        processUptimeSeconds:
          type: integer
          format: int64
          nullable: true
          description: 当前进程运行时长（秒，挂钟口径，仅本进程，不做系统进程枚举）。
        goroutineCount:
          type: integer
          format: int64
          nullable: true
        openFileDescriptors:
          type: integer
          format: int64
          nullable: true
    HostMonitoring:
      type: object
      required: [hostState, effectiveBucket, samples]
      properties:
        hostState:
          type: string
          enum: [healthy, stale, unknown]
        latestSampleAt:
          type: string
          format: date-time
        effectiveBucket:
          $ref: "#/components/schemas/ObservabilityBucket"
        latest:
          $ref: "#/components/schemas/HostMetricPoint"
        samples:
          type: array
          items:
            $ref: "#/components/schemas/HostMetricPoint"
...
    Error:
      type: object
      required: [error]
      properties:
        operationId:
          type: string
          description: 统一制品操作失败时返回的追踪标识
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...